feat: Add Docker config (#7)
* fix: Update timeConverter test to handle timezone changes

This commit updates the timeConverter tests and package.json script to account for timezone differences that occur when running tests in different environments.  Specifically, the timezone is now explicitly set to UTC to avoid test failures.  Additionally, the `timeConverter.js` file was renamed to `TimeConverter.js` to follow naming conventions.

* feat: Add Dockerfile and docker-compose

This commit introduces Dockerfile and a docker-compose.yml
file to enable packing the app as a docker image.

---------

Co-authored-by: JChris246 <JChris246@users.noreply.github.com>